requirements.txt
openai>=1.13.3
requests
pandas
python-dotenv
backoff
